Changed the calculation of the alert time to respect a reference day
The allert time (second/day/week) is now based on a reference time. This means:
nextAlertTime = referenceDay + (x * alert-intervall) so that x is as small as possible

Example:

The alert intervall is set to 2 days on monday: the next alert will pop up on wednesday.
If you forget to press the 'done' button on wednesday but press it on thursday the next alert will
nevertheless show on friday as it is calculated in respect to the reference day:

On thursday:
friday = monday + (2 * 2 days)